TKGAT finished

+++ main_ecfkg.py
@@ -1,5 +1,5 @@
 import os
-os.environ['CUDA_VISIBLE_DEVICES'] = '3'
+# os.environ['CUDA_VISIBLE_DEVICES'] = '3'
 
 import random
 import logging
@@ -21,6 +21,13 @@ from utility.metrics import *
 from utility.helper import *
 from utility.loader_ecfkg import DataLoaderECFKG
 
+def evaluate_with_scores(cf_scores, train_user_dict, test_user_dict, user_ids_batches, item_ids, K):
+    user_ids = np.concatenate(user_ids_batches)
+    precision_batch, recall_batch, ndcg_batch = calc_metrics_at_k(cf_scores, train_user_dict, test_user_dict, user_ids, item_ids, K)
+    recall = [np.mean(recall_batch[k]) for k in range(len(K))]
+    ndcg = [np.mean(ndcg_batch[k]) for k in range(len(K))]
+    precision = [np.mean(precision_batch[k]) for k in range(len(K))]
+    return precision, recall, ndcg
 
 def evaluate(model, user_ids_batches, item_ids, relation_u2i_id, train_user_dict, test_user_dict, K):
     model.eval()
@@ -33,6 +40,9 @@ def evaluate(model, user_ids_batches, item_ids, relation_u2i_id, train_user_dict
     precision = []
     recall = []
     ndcg = []
+    precision_k = []
+    recall_k = []
+    ndcg_k = []
 
     with torch.no_grad():
         for user_ids_batch in user_ids_batches:
@@ -48,9 +58,18 @@ def evaluate(model, user_ids_batches, item_ids, relation_u2i_id, train_user_dict
             ndcg.append(ndcg_batch)
 
     cf_scores = np.concatenate(cf_scores, axis=0)
-    precision_k = sum(np.concatenate(precision)) / n_users
-    recall_k = sum(np.concatenate(recall)) / n_users
-    ndcg_k = sum(np.concatenate(ndcg)) / n_users
+    if type(K) == int:
+        precision_k = np.mean(precision)
+        recall_k = np.mean(recall)
+        ndcg_k = np.mean(ndcg)
+    else:
+        for k in range(len(K)):
+            precision_k.append(np.mean(np.concatenate([precision[i][k] for i in range(len(precision))])))
+            recall_k.append(np.mean(np.concatenate([recall[i][k] for i in range(len(recall))])))
+            ndcg_k.append(np.mean(np.concatenate([ndcg[i][k] for i in range(len(ndcg))])))
+    # precision_k = sum(np.concatenate(precision)) / n_users
+    # recall_k = sum(np.concatenate(recall)) / n_users
+    # ndcg_k = sum(np.concatenate(ndcg)) / n_users
     return cf_scores, precision_k, recall_k, ndcg_k
 
 
@@ -66,7 +85,7 @@ def train(args):
 
     # GPU / CPU
     use_cuda = torch.cuda.is_available()
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    device = torch.device(f"cuda:{args.gpu}" if torch.cuda.is_available() else "cpu")
     n_gpu = torch.cuda.device_count()
     if n_gpu > 0:
         torch.cuda.manual_seed_all(args.seed)
@@ -100,8 +119,8 @@ def train(args):
         model = load_model(model, args.pretrain_model_path)
 
     model.to(device)
-    if n_gpu > 1:
-        model = nn.parallel.DistributedDataParallel(model)
+    # if n_gpu > 1:
+    #     model = nn.parallel.DistributedDataParallel(model)
     logging.info(model)
 
     optimizer = optim.Adam(model.parameters(), lr=args.lr)
@@ -145,7 +164,7 @@ def train(args):
         # evaluate cf
         if (epoch % args.evaluate_every) == 0:
             time1 = time()
-            _, precision, recall, ndcg = evaluate(model, user_ids_batches, item_ids, relation_u2i_id, data.train_user_dict, data.test_user_dict, args.K)
+            _, precision, recall, ndcg = evaluate(model, user_ids_batches, item_ids, relation_u2i_id, data.train_user_dict, data.valid_user_dict, args.K)
             logging.info('CF Evaluation: Epoch {:04d} | Total Time {:.1f}s | Precision {:.4f} Recall {:.4f} NDCG {:.4f}'.format(epoch, time() - time1, precision, recall, ndcg))
 
             epoch_list.append(epoch)
@@ -165,11 +184,16 @@ def train(args):
     # save model
     save_model(model, args.save_dir, epoch)
 
+    # test best model
+    best_model_dir = os.path.join(args.save_dir, 'model_epoch{}.pth'.format(best_epoch))
+    model = load_model(model, best_model_dir)
+    model.to(device)
+
     # save metrics
     _, precision, recall, ndcg = evaluate(model, user_ids_batches, item_ids, relation_u2i_id, data.train_user_dict, data.test_user_dict, args.K)
     logging.info('Final CF Evaluation: Precision {:.4f} Recall {:.4f} NDCG {:.4f}'.format(precision, recall, ndcg))
 
-    epoch_list.append(epoch)
+    epoch_list.append('Test_best')
     precision_list.append(precision)
     recall_list.append(recall)
     ndcg_list.append(ndcg)
@@ -189,38 +213,55 @@ def predict(args):
 
     # load data
     data = DataLoaderECFKG(args, logging)
+    K = np.arange(1,21)
 
-    user_ids = list(data.test_user_dict.keys())
-    user_ids_batches = [user_ids[i: i + args.test_batch_size] for i in range(0, len(user_ids), args.test_batch_size)]
-    user_ids_batches = [torch.LongTensor(d) for d in user_ids_batches]
-    if use_cuda:
-        user_ids_batches = [d.to(device) for d in user_ids_batches]
+    if not os.path.exists(os.path.join(args.save_dir, 'cf_scores.npy')):
+        user_ids = list(data.test_user_dict.keys())
+        user_ids_batches = [user_ids[i: i + args.test_batch_size] for i in range(0, len(user_ids), args.test_batch_size)]
+        user_ids_batches = [torch.LongTensor(d) for d in user_ids_batches]
+        if use_cuda:
+            user_ids_batches = [d.to(device) for d in user_ids_batches]
 
-    item_ids = torch.arange(data.n_items, dtype=torch.long)
-    if use_cuda:
-        item_ids = item_ids.to(device)
+        item_ids = torch.arange(data.n_items, dtype=torch.long)
+        if use_cuda:
+            item_ids = item_ids.to(device)
 
-    relation_u2i_id = torch.LongTensor([data.relation_u2i_id])
-    if use_cuda:
-        relation_u2i_id = relation_u2i_id.to(device)
+        relation_u2i_id = torch.LongTensor([data.relation_u2i_id])
+        if use_cuda:
+            relation_u2i_id = relation_u2i_id.to(device)
 
-    # load model
-    model = ECFKG(args, data.n_users, data.n_entities, data.n_relations)
-    model = load_model(model, args.pretrain_model_path)
-    model.to(device)
+        # load model
+        model = ECFKG(args, data.n_users, data.n_entities, data.n_relations)
+        model = load_model(model, get_best_model(args.save_dir))
+        print(f'Loaded {get_best_model(args.save_dir)}')
+        model.to(device)
 
-    # predict
-    cf_scores, precision, recall, ndcg = evaluate(model, user_ids_batches, item_ids, relation_u2i_id, data.train_user_dict, data.test_user_dict, args.K)
-    np.save(args.save_dir + 'cf_scores.npy', cf_scores)
-    print('CF Evaluation: Precision {:.4f} Recall {:.4f} NDCG {:.4f}'.format(precision, recall, ndcg))
+        # predict
+        cf_scores, precision, recall, ndcg = evaluate(model, user_ids_batches, item_ids, relation_u2i_id, data.train_user_dict, data.test_user_dict, K)
+        np.save(args.save_dir + 'cf_scores.npy', cf_scores)
+        # print('CF Evaluation: Precision {:.4f} Recall {:.4f} NDCG {:.4f}'.format(precision, recall, ndcg))
+    else:
+        user_ids = list(data.test_user_dict.keys())
+        user_ids_batches = [user_ids[i: i + args.test_batch_size] for i in range(0, len(user_ids), args.test_batch_size)]
+        user_ids_batches = [torch.LongTensor(d) for d in user_ids_batches]
+        item_ids = torch.arange(data.n_items, dtype=torch.long)
+        cf_scores = torch.tensor(np.load(os.path.join(args.save_dir, 'cf_scores.npy')))
+        precision, recall, ndcg = evaluate_with_scores(cf_scores, data.train_user_dict, data.test_user_dict, user_ids_batches, item_ids, K)
 
-
+    if not os.path.exists(args.result_dir):
+        os.makedirs(args.result_dir)
+    with open(os.path.join(args.result_dir, 'test_result.tsv'), mode='w') as f:
+        f.write('K\tprecision@K\trecall@K\tndcg@K\n')
+        for k in K:
+            f.write('{}\t{}\t{}\t{}\n'.format(k, precision[k-1], recall[k-1], ndcg[k-1]))
 
 if __name__ == '__main__':
     args = parse_ecfkg_args()
-    train(args)
-    # predict(args)
-
+    # os.environ['CUDA_VISIBLE_DEVICES'] = str(args.gpu)
+    if args.test:
+        predict(args)
+    else:
+        train(args)
